Move repo construction functions to RepoFactory
